+++ audio_be/src/commonMain/kotlin/voices/strip/pitch/PitchEnvelopeRenderer.kt
@@ -18,7 +18,8 @@ import kotlin.math.pow
  */
 class PitchEnvelopeRenderer(
     private val pitchEnvelope: Voice.PitchEnvelope,
-    private val startFrame: Int,
+        // Absolute backend frame — Double, see RenderClock.cursorFrame.
+    private val startFrame: Double,
 ) : BlockRenderer {
 
     override fun render(ctx: BlockContext) {
@@ -26,7 +27,7 @@ class PitchEnvelopeRenderer(
         val pEnv = pitchEnvelope
 
         // Compute voice-relative position as Int (once per block)
-        val blockRelStart = (ctx.blockStart + ctx.offset) - startFrame
+        val blockRelStart = (ctx.blockStart + ctx.offset - startFrame).toInt()
 
         if (ctx.freqModBufferWritten) {
             for (i in 0 until ctx.length) {
